feat: update view logs
